+++ src/components/3dModelComponent/index.js
@@ -1,7 +1,7 @@
 import { CustomButton } from "./CustomButton";
-import { GaleryPicker } from "./GaleryPicker";
 import { ColorPicker } from "./ColorPicker";
 import { FilePicker } from "./FilePicker";
 import { Tab } from "./Tab";
+import { CanvasModel } from "./index.jsx";
 
-export { CustomButton, GaleryPicker, ColorPicker, FilePicker, Tab };
+export { CustomButton, ColorPicker, FilePicker, Tab, CanvasModel };