+++ src/components/3dModelComponent/index.jsx
@@ -13,7 +13,6 @@ export const CanvasModel = () => {
         gl={{ preserveDrawingBuffer: true }}
         className="w-full max-w-full transition-all ease-in"
       >
-        <ambientLight intensity={0.5} />
         <Environment preset="city" />
 
         <CameraRig>