+++ src/components/orgamisms/DesignComponent.jsx
@@ -1,12 +1,10 @@
-import { Customiser } from "../pages/Customiser";
-import { CanvasModel } from "../canvasModel";
-import { HomePageCustom } from "../pages/HomePageCustom";
-
+import { Customiser } from "../../pages/Customiser";
+import { CanvasModel } from "../3dModelComponent";
+import { HomePageCustom } from "../../pages/HomePageCustom";
 export const DesignComponent = () => {
   return (
     <div className="flex flex-col justify-start h-screen  md:w-full  mx-auto ">
       <HomePageCustom />
-
       <CanvasModel />
       <Customiser />
     </div>