+++ src/components/orgamisms/Header.jsx
@@ -33,21 +33,41 @@ export const Header = () => {
           <Logo />
         </Link>
         <div className="sm:order-3 flex items-center gap-x-2">
-          <button
+          <Button
             type="button"
-            className="sm:hidden hs-collapse-toggle p-2.5 inline-flex justify-center items-center gap-x-2 rounded-lg border border-gray-200 bg-white text-gray-800 shadow-sm hover:bg-gray-50 disabled:opacity-50 disabled:pointer-events-none dark:bg-transparent dark:border-gray-700 dark:text-white dark:hover:bg-white/10 dark:focus:outline-none dark:focus:ring-1 dark:focus:ring-gray-600"
+            className="sm:hidden hs-collapse-toggle p-2.5 inline-flex justify-center items-center gap-x-2 rounded-lg border border-gray-200 bg-white text-gray-800 shadow-sm disabled:opacity-50 disabled:pointer-events-none dark:bg-transparent dark:border-gray-700 dark:text-white dark:focus:outline-none dark:focus:ring-1 dark:focus:ring-gray-600"
             onClick={toggleCollapse}
             aria-controls="navbar-alignment"
             aria-label="Toggle navigation"
           >
-            {/* SVG code here */}
-          </button>
+            <svg
+              className="hs-collapse-open:hidden w-4 h-4"
+              width="16"
+              height="16"
+              fill="currentColor"
+              viewBox="0 0 16 16"
+            >
+              <path
+                fillRule="evenodd"
+                d="M2.5 12a.5.5 0 0 1 .5-.5h10a.5.5 0 0 1 0 1H3a.5.5 0 0 1-.5-.5zm0-4a.5.5 0 0 1 .5-.5h10a.5.5 0 0 1 0 1H3a.5.5 0 0 1-.5-.5zm0-4a.5.5 0 0 1 .5-.5h10a.5.5 0 0 1 0 1H3a.5.5 0 0 1-.5-.5z"
+              />
+            </svg>
+            <svg
+              className="hs-collapse-open:block hidden w-4 h-4"
+              width="16"
+              height="16"
+              fill="currentColor"
+              viewBox="0 0 16 16"
+            >
+              <path d="M4.646 4.646a.5.5 0 0 1 .708 0L8 7.293l2.646-2.647a.5.5 0 0 1 .708.708L8.707 8l2.647 2.646a.5.5 0 0 1-.708.708L8 8.707l-2.646 2.647a.5.5 0 0 1-.708-.708L7.293 8 4.646 5.354a.5.5 0 0 1 0-.708z" />
+            </svg>
+          </Button>
           {userLoggedIn ? (
             <UserMenu onLogout={handleLogout} />
           ) : (
             <>
               <Link to="/login">
-                <Button>Log In</Button>
+                <Button className="hover:bg-DarkBlue">Log In</Button>
               </Link>
               <Link to="/signup">
                 <Button className="hover:bg-SuperPink">Sign Up</Button>