+++ src/components/orgamisms/MainArticle.jsx
@@ -1,4 +1,4 @@
-import { DesignComponent } from "../pages/DesignComponent";
+import { DesignComponent } from "./DesignComponent";
 export const MainArticle = () => {
   return (
     <section className="container mx-auto mb-5">